完善权限,抽奖活动

+++ app/Http/Controllers/Main/ActivityController.php
@@ -6,6 +6,7 @@ use App\Activity;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 
 class ActivityController extends Controller
@@ -18,10 +19,16 @@ class ActivityController extends Controller
     }
     //创建活动
     public function create(){
+        if(!Auth::user()->can('activity.create')){
+            return view('Error.nopage');
+        }
         return view('activity.create');
     }
     //保存
     public function store(Request $request){
+        if(!Auth::user()->can('activity.create')){
+            return view('Error.nopage');
+        }
         $this->validate($request,
             [
                 'name'=>'required|unique:activities',
@@ -64,20 +71,32 @@ class ActivityController extends Controller
     }
     //列表
     public function index(){
+        if(!Auth::user()->can('activity.index')){
+            return view('Error.nopage');
+        }
         $now = Carbon::now();
         $activitys = Activity::all();
         return view('activity.index',compact('activitys','now'));
     }
     //查看
     public function show(Activity $activity){
+        if(!Auth::user()->can('activity.show')){
+            return view('Error.nopage');
+        }
         return view('activity.show',compact('activity'));
     }
     //活动-编辑
     public function edit(Activity $activity){
+        if(!Auth::user()->can('activity.edit')){
+            return view('Error.nopage');
+        }
         return view('activity.edit',compact('activity'));
     }
     //编辑-保存
     public function update(Request $request,Activity $activity){
+        if(!Auth::user()->can('activity.edit')){
+            return view('Error.nopage');
+        }
         $this->validate($request,
             [
                 'name'=>[
@@ -121,6 +140,9 @@ class ActivityController extends Controller
     }
     //删除
     public function destroy(Activity $activity){
+        if(!Auth::user()->can('activity.destroy')){
+            return view('Error.nopage');
+        }
         $activity->delete();
         echo "success";
     }
